Rename date time picker to time range picker.
