+++ graylog2-web-interface/src/views/components/searchbar/time-range-picker/RelativeRangeSelect.tsx
@@ -21,11 +21,11 @@ import moment from 'moment';
 
 import { RELATIVE_RANGE_TYPES } from 'views/Constants';
 import { Select } from 'components/common';
-import type { RangeClassified } from 'views/components/searchbar/date-time-picker/types';
+import type { RangeClassified } from 'views/components/searchbar/time-range-picker/types';
 
 import { isTypeRelativeClassified, RELATIVE_CLASSIFIED_ALL_TIME_RANGE } from './RelativeTimeRangeClassifiedHelper';
 import RelativeRangeValueInput from './RelativeRangeValueInput';
-import type { TimeRangeDropDownFormValues } from './TimeRangeDropdown';
+import type { TimeRangeDropDownFormValues } from './TimeRangePicker';
 
 const RangeWrapper = styled.div`
   flex: 4;
