+++ graylog2-web-interface/src/views/components/searchbar/time-range-picker/RelativeTimeRangeClassifiedHelper.ts
@@ -23,7 +23,7 @@ import type {
   KeywordTimeRange,
   NoTimeRangeOverride, TimeRange,
 } from 'views/logic/queries/Query';
-import type { RelativeTimeRangeClassified, RangeClassified } from 'views/components/searchbar/date-time-picker/types';
+import type { RelativeTimeRangeClassified, RangeClassified } from 'views/components/searchbar/time-range-picker/types';
 import { isTypeRelativeWithStartOnly } from 'views/typeGuards/timeRange';
 
 const DEFAULT_CLASSIFIED_RANGE: RangeClassified = {
