+++ graylog2-web-interface/src/views/components/searchbar/time-range-picker/TabAbsoluteTimeRange.tsx
@@ -26,11 +26,11 @@ import type { AbsoluteTimeRange } from 'views/logic/queries/Query';
 import useUserDateTime from 'hooks/useUserDateTime';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
 import useSearchConfiguration from 'hooks/useSearchConfiguration';
-import TabPresetDropdown from 'views/components/searchbar/date-time-picker/TabPresetDropdown';
+import TabPresetDropdown from 'views/components/searchbar/time-range-picker/TabPresetDropdown';
 
 import AbsoluteCalendar from './AbsoluteCalendar';
 import AbsoluteTimestamp from './AbsoluteTimestamp';
-import type { TimeRangeDropDownFormValues } from './TimeRangeDropdown';
+import type { TimeRangeDropDownFormValues } from './TimeRangePicker';
 
 type Props = {
   disabled: boolean,
