+++ graylog2-web-interface/src/views/components/searchbar/time-range-picker/TabDisabledTimeRange.tsx
